:truck: [wallet/test] Move address test to separate file
We can use one test file per interface.
